[DB] project two - implement basic sql function
1. implement create,insert,update,delete,drop,show tables function.
2. develop B+tree data structure to store data.
3. use special tables to record user tables info.

+++ database_design/dbLite/src/edu/utd/db/zhySql/Page/cell.java
@@ -9,12 +9,12 @@ package edu.utd.db.zhySql.Page;
 public class cell {
 	//int pageNumber;
 	int rowId;
-	long location; //absolute address
+	short location; //absolute address
 	
-	public long getLocation() {
+	public short getLocation() {
 		return location;
 	}
-	public void setLocation(long location) {
+	public void setLocation(short location) {
 		this.location = location;
 	}
